+++ kylin-dorm/src/main/java/com/store59/kylin/dorm/service/DormService.java
@@ -15,7 +15,7 @@ public class DormService {
 		return dormDao.selectByPrimaryKey(dormId);
 	}
 
-	public Dorm getDormByUid(Integer uid) {
+	public Dorm getDormByUid(Long uid) {
 		return dormDao.selectByUid(uid);
 	}
